requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asttokens==2.0.5
async-lru==2.0.4
attrs==23.1.0
Babel==2.11.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.2
bleach==4.1.0
blinker==1.8.2
Brotli==1.0.9
build==1.2.1
cachetools==5.3.3
camelot-py==0.9.0
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==2.0.4
chroma-hnswlib==0.7.3
chromadb==0.5.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.1
contourpy==1.2.1
cryptography==42.0.7
cycler==0.12.1
dataclasses-json==0.6.6
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
dnspython==2.6.1
email_validator==2.1.1
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
Faker==25.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.16.2
favicon==0.7.0
filelock==3.14.0
flatbuffers==24.3.25
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.29.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.64.0
h11==0.14.0
htbuilder==0.6.2
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.1
humanfriendly==10.0
idna==3.7
importlib-metadata==7.0.0
importlib_resources==6.4.0
InstructorEmbedding==1.0.1
intel-openmp==2021.4.0
ipykernel==6.28.0
ipython==8.20.0
ipywidgets==8.1.2
jedi==0.18.1
Jinja2==3.1.3
joblib==1.4.2
json5==0.9.6
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter_client==8.6.0
jupyter-console==6.6.3
jupyter_core==5.5.0
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_server==2.10.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.11
jupyterlab-pygments==0.1.2
jupyterlab_server==2.25.1
jupyterlab-widgets==3.0.10
kiwisolver==1.4.5
kubernetes==29.0.0
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.1
langchain-text-splitters==0.2.0
langsmith==0.1.62
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.3
marshmallow==3.21.2
matplotlib==3.9.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.4
mkl==2021.4.0
mmh3==4.1.0
monotonic==1.6
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
notebook==7.0.8
notebook_shim==0.2.3
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.18.0
openai==1.30.2
opencv-python==4.9.0.80
openpyxl==3.1.2
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.10.3
overrides==7.4.0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.0
parso==0.8.3
pdfminer.six==20231228
pillow==10.3.0
pip==24.0
platformdirs==3.10.0
ply==3.11
posthog==3.5.0
prometheus-client==0.14.1
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.0
pure-eval==0.2.2
pyarrow==16.1.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.1
Pygments==2.15.1
pymdown-extensions==10.8.1
pyparsing==3.1.2
PyPDF2==2.9.0
PyPika==0.48.9
pyproject_hooks==1.1.0
PyQt5==5.15.10
PyQt5-sip==12.13.0
pyreadline3==3.4.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
pytz==2024.1
pywin32==305.1
pywinpty==2.0.10
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.30.2
regex==2024.5.15
requests==2.31.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.10.6
rsa==4.9
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.2.0
setuptools==69.5.1
shellingham==1.5.4
sip==6.7.12
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.30
st-annotated-text==4.0.1
stack-data==0.2.0
starlette==0.37.2
streamlit==1.35.0
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.4.2
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.4
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
sympy==1.12
tbb==2021.12.0
tenacity==8.3.0
terminado==0.17.1
threadpoolctl==3.5.0
tinycss2==1.2.1
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.3.0
torchvision==0.18.0
tornado==6.3.3
tqdm==4.66.4
traitlets==5.7.1
transformers==4.41.1
typer==0.12.3
typing_extensions==4.11.0
typing-inspect==0.9.0
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.29.0
validators==0.28.2
watchdog==4.0.1
watchfiles==0.21.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
wheel==0.43.0
widgetsnbextension==4.0.10
win-inet-pton==1.1.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.18.2
